Move convertDescriptorToString to jest-util
