+++ packages/jest-circus/package.json
@@ -13,7 +13,8 @@
     "jest-diff": "^23.0.0",
     "jest-matcher-utils": "^23.0.0",
     "jest-message-util": "^23.0.0",
-    "jest-snapshot": "^23.0.0"
+    "jest-snapshot": "^23.0.0",
+    "jest-util": "^23.0.0"
   },
   "devDependencies": {
     "jest-runtime": "^23.0.0"
