+++ packages/jest-circus/src/utils.js
@@ -21,6 +21,7 @@ import type {
   TestName,
   TestResults,
 } from 'types/Circus';
+import {convertDescriptorToString} from 'jest-util';
 
 export const makeDescribe = (
   name: BlockName,
@@ -270,33 +271,3 @@ export const invariant = (condition: *, message: string) => {
     throw new Error(message);
   }
 };
-
-// See: https://github.com/facebook/jest/pull/5154
-function convertDescriptorToString(descriptor) {
-  if (
-    typeof descriptor === 'string' ||
-    typeof descriptor === 'number' ||
-    descriptor === undefined
-  ) {
-    return descriptor;
-  }
-
-  if (typeof descriptor !== 'function') {
-    throw new Error('describe expects a class, function, number, or string.');
-  }
-
-  if (descriptor.name !== undefined) {
-    return descriptor.name;
-  }
-
-  // Fallback for old browsers, pardon Flow
-  const stringified = descriptor.toString();
-  const typeDescriptorMatch = stringified.match(/class|function/);
-  const indexOfNameSpace =
-    // $FlowFixMe
-    typeDescriptorMatch.index + typeDescriptorMatch[0].length;
-  // $FlowFixMe
-  const indexOfNameAfterSpace = stringified.search(/\(|\{/, indexOfNameSpace);
-  const name = stringified.substring(indexOfNameSpace, indexOfNameAfterSpace);
-  return name.trim();
-}
